skin-cancer-system/mlservice/requirements.txt
flask
flask-cors
tensorflow
numpy
opencv-python
pillow
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
